+++ src/Ext/WarheadType/Body.cpp
@@ -298,8 +298,6 @@ void WarheadTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->AffectsNeutral.Read(exINI, pSection, "AffectsNeutral");
 	this->HealthCheck = this->AffectsBelowPercent > 0.0 || this->AffectsAbovePercent < 1.0;
 
-	this->OnlyAttacker.Read(exINI, pSection, "OnlyAttacker");
-
 	if (this->AffectsAbovePercent > this->AffectsBelowPercent)
 		Debug::Log("[Developer warning][%s] AffectsAbovePercent is bigger than AffectsBelowPercent, the warhead will never activate!\n", pSection);
 
@@ -529,8 +527,6 @@ void WarheadTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->AffectsAbovePercent)
 		.Process(this->AffectsNeutral)
 		.Process(this->HealthCheck)
-
-		.Process(this->OnlyAttacker)
 
 		.Process(this->InflictLocomotor)
 		.Process(this->RemoveInflictedLocomotor)