+++ src/Ext/WarheadType/Body.h
@@ -187,8 +187,6 @@ public:
 		Valueable<double> AffectsAbovePercent;
 		Valueable<bool> AffectsNeutral;
 
-		Valueable<bool> OnlyAttacker;
-
 		// Ares tags
 		// http://ares-developers.github.io/Ares-docs/new/warheads/general.html
 		Valueable<bool> AffectsEnemies;
@@ -374,8 +372,6 @@ public:
 			, AffectsBelowPercent { 1.0 }
 			, AffectsAbovePercent { 0.0 }
 			, AffectsNeutral { true }
-
-			, OnlyAttacker { false }
 
 			, AffectsEnemies { true }
 			, AffectsOwner {}