+++ docs/Whats-New.md
@@ -428,7 +428,7 @@ New:
 - [Units can customize the attack voice that plays when using more weapons](New-or-Enhanced-Logics.md#multi-voiceattack) (by FlyStar)
 - Customize squid grapple animation (by NetsuNegi)
 - [Auto deploy for GI-like infantry](Fixed-or-Improved-Logics.md#auto-deploy-for-gi-like-infantry) (by TaranDahl)
-- When a warhead has `OnlyAttacker=yes`, it prevents other units using that warhead from attacking the same target (by FlyStar)
+- When a weapon has `OnlyAttacker=yes`, it prevents other units using that weapon from attacking the same target (by FlyStar)
 
 Vanilla fixes:
 - Fixed sidebar not updating queued unit numbers when adding or removing units when the production is on hold (by CrimRecya)